Update README.md
ROAD SIGN RECOGNITION

ABSTRACT:
Road accidents are mostly caused by carelessness of the driver. Several factors like alcohol drinking, drowsiness, lack of concentration, etc. affect the awareness of drivers to traffic signs including visual complexity, environmental condition, and poor driving knowledge. Detection and recognition of road signs has long been in the center of interest for having great impact on the safety of the driver. However, it remains some problems. First, it is difficult to track an object in an image sequence captured from a vehicle undergoing a generally non uniform motion. Second, air pollution and weather conditions may decrease the visibility of road signs, besides that, outdoor lighting conditions vary from day to night and may affect the apparent colors of road signs. Finally, recognition system has to be fast and robust which increase the difficulty of the automatic detection of the road signs. For all these reasons, reliable detection system of road signs from various scenes becomes rather challenging. The classification module determines the type of detected road signs using a scikit-learn. when the road signs are identified manually by the driver on their own. The disadvantage is that the driver may miss out some signs by mistake. If signs like speed breaker are missed out, it may lead to accidents. a computer vision system which is able to perform automated recognition that could be used to notify a human driver to the presence and nature of road signs is developed. The presence of road sign is identified using deep learning and the nature of road sign is identified through the same.